requirements.txt
python-dotenv
# wikipedia
# tiktoken
neo4j
# streamlit
# Pillow
fastapi
# PyPDF2
pydantic
uvicorn
# sse-starlette
# boto3
# streamlit==1.32.1
# missing from the langchain base image?
langchain-openai==0.3.8
# langchain-community==0.3.19
# langchain-google-genai==2.0.11
# langchain-ollama==0.2.3
# langchain-huggingface==0.1.2
# langchain-aws==0.2.15
langchain-neo4j==0.4.0
numpy

# Authentication backend dependencies
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
